removed main and renamed guibuild to main
